// src/store/enum/populateExcel.enum.ts
export enum EGender {
  MALE = "Male",
  FEMALE = "Female",
  THIRD = "Third Gender",
}

export enum ESocialCategory {
  UNRESERVED = "Unreserved",
  SC = "SC",
  ST = "ST",
  BC = "BC",
  MBC = "MBC",
}

export enum EReligion {
  HINDU = "Hindu",
  ISLAM = "Islam",
  CHRISTIAN = "Christian",
  SIKH = "Sikh",
  BUDDHIST = "Buddhist",
  PARSI = "Parsi",
  JAIN = "Jain",
  OTHERS = "Others",
}

export enum ELanguage {
  Assamese = "Assamese",
  Bengali = "Bengali",
  Gujarati = "Gujarati",
  Hindi = "Hindi",
  Kannada = "Kannada",
  Kashmiri = "Kashmiri",
  Konkani = "Konkani",
  Manipuri = "Manipuri",
  Marathi = "Marathi",
  Nepali = "Nepali",
  Odia = "Odia",
  Punjabi = "Punjabi",
  Sanskrit = "Sanskrit",
  Sindhi = "Sindhi",
  Tamil = "Tamil",
  Telugu = "Telugu",
  Urdu = "Urdu",
  English = "English",
  Bodo = "Bodo",
  Dogri = "Dogri",
  Khasi = "Khasi",
  Garo = "Garo",
  Mizo = "Mizo",
  Bhutia = "Bhutia",
  Lepcha = "Lepcha",
  Limboo = "Limboo",
  French = "French",
  Santhali = "Santhali",
  Maithali = "Maithali",
  Otherlanguages = "Other languages",
}

export enum EPreviousToClass1 {
  ANOTHER = "Another School",
  ANGANWADI = "Anganwadi ECCE",
  NA = "NA",
}

export enum EClothSet {
  NONE = "None",
  ONE = "One Set",
  TWO = "Two Set",
  NA = "NA",
}

export enum EDisability {
  NA = "NA",
  BLINDNESS = "Blindness",
  LOW_VISION = "Low-Vision",
  HEARING = "Hearing Impairment(deaf and hard to hearing)",
  SPEECH = "Speech and Language disablity",
  LOCOMOTOR = "Locomotor Disability",
  MENTAL = "Mental Illness",
  LEARNING = "Specific Learning Disablities",
  CEREBRAL = "Cerebral Palsy",
  AUTISM = "Autism Spectrum Disorder",
  DEAF_BLINDNESS = "Multiple Disablities including deaf blindness",
  LEPROSY = "Leprosy Cured persons",
  DWARIFISM = "Dwarifism",
  INTELLECTUAL = "Intellectual Disablity",
  DYSTROPHY = "Muscular Dystrophy",
  NUROLOGICAL = "Chronic Nurological conditions",
  MULTIPLE = "Multiple Sclerosis",
  THALASSEMIA = "Thalassemia",
  HEMOPHILIA = "Hemophilia",
  SICKLE_CELL = "Sickle Cell disease",
  ACID_ATTACK = "Acid Attack Victim",
  PARKINSON = "Parkinson's disease",
}

export enum EFacilities {
  NA = "NA",
  BRAILLE_BOOKS = "Braille Books",
  BRAILLE_KIT = "Braille Kit",
  LOW_VISION_KIT = "Low Vision Kit",
  HEARING_AID = "Hearing aid",
  BRACES = "Braces",
  CRUTCHES = "Crutches",
  WHEEL_CHAIR = "Wheel Chair",
  TRICYCLE = "Tricycle",
  CALIPER = "Caliper",
  OTHERS = "Others",
}

export enum EFreeBooks {
  YES = "Yes",
  NO = "No",
  NA = "NA",
}

export enum EFreeTransport {
  YES = "Yes",
  NO = "No",
  NA = "NA",
}

export enum EFreeEscort {
  YES = "Yes",
  NO = "No",
  NA = "NA",
}

export enum EFreeBicycle {
  YES = "Yes",
  NO = "No",
  NA = "NA",
}

export enum EFreeHostel {
  NA = "NA",
  KGBV = "KGBV",
  JBAV = "JBAV",
  OTHERS = "Others",
}

export enum ESpecialTraining {
  RESIDENTIAL = "Residential",
  NON_RESIDENTIAL = "Non-Residential",
  BOTH = "Both Residential And Non-Residential",
  NA = "NA",
}

export enum EHomeless {
  YES = "Yes",
  NO = "No",
  NA = "NA",
}

export enum EFreeEducation {
  YES = "Yes",
  NO = "No",
  NA = "NA",
}

export enum ENewAdmission {
  NEW = "New",
  SAME = "Same School",
}

export enum EBPL {
  YES = "Yes",
  NO = "No",
}

export enum EDisadvantageGroup {
  YES = "Yes",
  NO = "No",
}

export enum EClass {
  LKG = "Lkg",
  UKG = "Ukg",
  Class1 = "Class 1",
  Class2 = "Class 2",
  Class3 = "Class 3",
  Class4 = "Class 4",
  Class5 = "Class 5",
  Class6 = "Class 6",
  Class7 = "Class 7",
  Class8 = "Class 8",
  NA = "NA",
}